Refactoring.

+++ src/com/zilen/messagingService/entity/attachment/Document.java
@@ -15,5 +15,5 @@ public class Document implements Attachment {
     @NonNull
     private String title;
     private String format;
-    private Byte size;
+    private int size;
 }
